refactor createTags.js functions
